03.05_Obtendo_Valor_Comissao_Servico

+++ api/models.py
@@ -68,8 +68,10 @@ class Diaria(models.Model):
     tempo_atendimento = models.IntegerField(null=False, blank=False)
     status = models.IntegerField(null=False, blank=False, choices=STATUS_DIARIA_CHOICES,
     default=1)
-    preco = models.FloatField(null=False, blank=False)
-    valor_comissao = models.FloatField(null=False, blank=False)
+    preco = models.DecimalField(null=False, blank=False, 
+    decimal_places=2, max_digits=5)
+    valor_comissao = models.DecimalField(null=False, blank=False, 
+    decimal_places=2, max_digits=5)    
     logradouro = models.CharField(max_length=60, null=False, blank=False)
     numero = models.CharField(max_length=10, null=False, blank=False)
     bairro = models.CharField(max_length=30, null=False, blank=False)
